Format JavaDocs, utilize new Storage exceptions, refactor Command execute functionality
