+++ src/main/java/duke/command/ListCommand.java
@@ -16,14 +16,14 @@ public class ListCommand extends Command {
 
     /**
      * Prints out list of tasks.
-     *  @param tasks list of tasks
-     * @param ui prints information to user
+     *
+     * @param tasks   list of tasks
+     * @param ui      prints information to user
      * @param storage manages user save file
      * @return TaskList required for indicating updating of tasks
      */
     @Override
     public Optional<TaskList> execute(TaskList tasks, Ui ui, Storage storage) {
-        // Print out all tasks
         ui.listTasks(tasks, "Here are the tasks in your list:");
         return Optional.empty();
     }
