+++ src/main/java/duke/exception/DukeStorageLoadException.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.exception;
 
 public class DukeStorageLoadException extends DukeException {
     /**
